chore: move back consensus' Worker under main consensus module.
